MP3 conversion working

+++ AudioUtilities/AudioFileConverter.swift
@@ -67,57 +67,30 @@ class AudioFileConverter{
         }
         
     }
-    class func convertToMP3(sourceURL: URL, destinationURL: URL, deleteSource: Bool){ //WIP
-           let asset = AVAsset(url: sourceURL)
-           do {
-               let reader = try AVAssetReader(asset: asset)
-               let track = asset .tracks(withMediaType: AVMediaType.audio)[0]
-               let compressionSettings : [String: Any] = [
-                   AVFormatIDKey: kAudioFormatMPEGLayer3,
-                   AVSampleRateKey : 48000//,
-//                   AVLinearPCMBitDepthKey : 16,
-//                   AVLinearPCMIsNonInterleaved : false,
-//                   AVLinearPCMIsFloatKey : false,
-//                   AVLinearPCMIsBigEndianKey : false,
-//                   AVNumberOfChannelsKey : 2
-               ]
-               let readerOutput = AVAssetReaderTrackOutput(track: track, outputSettings: compressionSettings)
-               reader.add(readerOutput)
-               let outURL = destinationURL
-               if FileManager.default.fileExists(atPath: outURL.relativeString){
-                   print("File exists! Kill!\(outURL.absoluteString)")
-               }
-               
-               let writer = try AVAssetWriter(outputURL: outURL, fileType: AVFileType.wav)
-               //Need to delete output file if it exists
-               let writerInput = AVAssetWriterInput(mediaType: AVMediaType.audio, outputSettings: compressionSettings)
-               writer.add(writerInput)
-               writer.startWriting()
-               reader.startReading()
-               let startTime = CMTime(seconds: 0, preferredTimescale: 1)
-               writer.startSession(atSourceTime: startTime)
-               var sample = readerOutput.copyNextSampleBuffer()
-               while sample != nil{
-                   while true {
-                       if writerInput.isReadyForMoreMediaData {
-                           writerInput.append(sample!)
-                           break
-                       }
-                   }
-                   sample = readerOutput.copyNextSampleBuffer()
-               }
-               writer.finishWriting { 
-                   print("status = \(writer.status)")
-                   print("Error = \(writer.error)")
-                   print("Done!")
-               }
-               if deleteSource{
-                   let fileManager = FileManager()
-                   try fileManager.removeItem(at: sourceURL)
-               }
-           } catch {
-               print(error)
-           }
-           
-       }
+    class func convertToMP3(sourceURL: URL, destinationURL: URL, deleteSource: Bool){
+        let task = Process()
+        let pipe = Pipe()
+        task.standardOutput = pipe
+        task.launchPath = "/usr/local/bin/lame"
+        let sourcePath = sourceURL.path
+        let destPath = destinationURL.path
+        task.arguments = [sourcePath, destPath]
+        task.launch()
+        let handle = pipe.fileHandleForReading
+        let data = handle.readDataToEndOfFile()
+        guard let result_s = NSString(data: data, encoding: String.Encoding.utf8.rawValue) else {
+            print("Couldn't get mp3 conversion results")
+            return
+        }
+        print(result_s)
+        if deleteSource{
+            let fileManager = FileManager()
+            do {
+                try fileManager.removeItem(at: sourceURL)
+            } catch {
+                print(error)
+            }
+        }
+        
+    }
 }
